Performance data generation
1. Added class for running performance tests
2. Added recursive task class for executing queries parallel mode
3. Added entity for storing user data for data generation
4. Added query for granting create table privilege
5. Added query for creating database with fluent size
6. Move test constants to separate static class
7. Added static class with common methods for Test data helpers
