+++ integ-tests/pom.xml
@@ -121,7 +121,7 @@
                 </executions>
                 <configuration>
                     <includePluginDependencies>true</includePluginDependencies>
-                    <mainClass>com.google.testdata.TestDataGenerator</mainClass>
+                    <mainClass>com.google.testdata.FunctionalTestDataGenerator</mainClass>
                 </configuration>
                 <dependencies>
                     <dependency>
