+++ integ-tests/src/test/java/com/google/integration/UsersTest.java
@@ -18,6 +18,12 @@ package com.google.integration;
 import static com.google.avro.AvroHelper.extractAvroDataFromFile;
 import static com.google.avro.AvroHelper.getLongNotNull;
 import static com.google.avro.AvroHelper.getStringNotNull;
+import static com.google.base.TestConstants.DB_NAME;
+import static com.google.base.TestConstants.ET_OUTPUT_PATH;
+import static com.google.base.TestConstants.PASSWORD_DB;
+import static com.google.base.TestConstants.SQL_REQUESTS_BASE_PATH;
+import static com.google.base.TestConstants.URL_DB;
+import static com.google.base.TestConstants.USERNAME_DB;
 import static com.google.tdjdbc.JdbcHelper.getStringNotNull;
 import static com.google.tdjdbc.JdbcHelper.getTimestampNotNull;
 import static java.lang.String.format;
